update financy statistics

+++ controllers/DiscountController.js
@@ -1,5 +1,13 @@
 const DiscountServices = require("../models/services/DiscountServices");
 exports.index = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     const discounts = await DiscountServices.getDiscountList();
     await res.render("pages/admin/discount", {
         discounts: discounts,
@@ -7,13 +15,37 @@ exports.index = async (req, res, next) => {
     });
 };
 exports.getAdd = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     res.render("pages/admin/discountDetail", { page: "discount" });
 };
 exports.postAdd = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     await DiscountServices.addDiscount(req.body);
     res.json({ log: "success" });
 };
 exports.getEdit = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     const discount = await DiscountServices.getDiscountByID(req.params.id);
     res.render("pages/admin/discountDetail", {
         discount: discount,
@@ -21,6 +53,14 @@ exports.getEdit = async (req, res, next) => {
     });
 };
 exports.postEdit = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     await DiscountServices.updateDiscount(
         req.params.id,
         req.body
@@ -28,6 +68,14 @@ exports.postEdit = async (req, res, next) => {
     res.json({ log: "success" });
 };
 exports.remove = async (req, res, next) => {
+    if (!req.user) {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
+    if (req.user.role != "manager") {
+        res.status(401);
+        return res.json({ error: "access denied" });
+    }
     await DiscountServices.deleteDiscount(req.params.id);
     res.json({ log: "success" });
 };
